Add Breakpoint class
Take the generic parts of the MspBreakpoint implementation
and put them in a Breakpoint class.

Rename XML config entry "contikicode" to "sourcecode"
since it is generic, but continue to recognize
the previous name in config files.

This will enable more code sharing for WatchpointMote.
