+++ java/org/contikios/cooja/mspmote/plugins/MspBreakpoint.java
@@ -30,17 +30,8 @@
 
 package org.contikios.cooja.mspmote.plugins;
 
-import java.awt.Color;
 import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
-
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
-import org.jdom2.Element;
-
-import org.contikios.cooja.Watchpoint;
+import org.contikios.cooja.Breakpoint;
 import org.contikios.cooja.mspmote.MspMote;
 import org.contikios.cooja.util.StringUtils;
 import se.sics.mspsim.core.Memory;
@@ -51,97 +42,22 @@ import se.sics.mspsim.core.MemoryMonitor;
  *
  * @author Fredrik Osterlind
  */
-public class MspBreakpoint implements Watchpoint {
-  private static final Logger logger = LogManager.getLogger(MspBreakpoint.class);
-
+public class MspBreakpoint extends Breakpoint {
   private final MspMote mspMote;
-
-  private long address = -1; /* Binary address */
-  private File codeFile = null; /* Source code, may be null*/
-  private int lineNr = -1; /* Source code line number, may be null */
-
   private MemoryMonitor memoryMonitor = null;
-
-  private boolean stopsSimulation = true;
-
-  private String msg = null;
-  private Color color = Color.BLACK;
-
-  private String contikiCode = null;
-
   public MspBreakpoint(MspMote mote) {
-    this.mspMote = mote;
+    super(mote);
+    mspMote = mote;
     /* expects setConfigXML(..) */
   }
 
   public MspBreakpoint(MspMote mote, long address, File codeFile, Integer lineNr) {
-    this(mote);
-    this.address = address;
-    this.codeFile = codeFile;
-    this.lineNr = lineNr;
-
-    createMonitor();
+    super(mote, address, codeFile, lineNr);
+    mspMote = mote;
   }
 
   @Override
-  public MspMote getMote() {
-    return mspMote;
-  }
-
-  @Override
-  public Color getColor() {
-    return color;
-  }
-  @Override
-  public void setColor(Color color) {
-    this.color = color;
-  }
-
-  @Override
-  public String getDescription() {
-    String desc = "";
-    if (codeFile != null) {
-      desc += codeFile.getPath() + ":" + lineNr + " (0x" + Long.toHexString(address) + ")";
-    } else if (address >= 0) {
-      desc += "0x" + Long.toHexString(address);
-    }
-    if (msg != null) {
-      desc += "\n\n" + msg;
-    }
-    return desc;
-  }
-  @Override
-  public void setUserMessage(String msg) {
-    this.msg = msg;
-  }
-  @Override
-  public String getUserMessage() {
-    return msg;
-  }
-
-  @Override
-  public File getCodeFile() {
-    return codeFile;
-  }
-  @Override
-  public int getLineNumber() {
-    return lineNr;
-  }
-  @Override
-  public long getExecutableAddress() {
-    return address;
-  }
-
-  @Override
-  public void setStopsSimulation(boolean stops) {
-    stopsSimulation = stops;
-  }
-  @Override
-  public boolean stopsSimulation() {
-    return stopsSimulation;
-  }
-
-  private void createMonitor() {
+  protected void createMonitor() {
     memoryMonitor = new MemoryMonitor.Adapter() {
       @Override
       public void notifyReadBefore(int addr, Memory.AccessMode mode, Memory.AccessType type) {
@@ -155,13 +71,13 @@ public class MspBreakpoint implements Watchpoint {
     mspMote.getCPU().addWatchPoint((int) address, memoryMonitor);
 
 
-    /* Remember Contiki code, to verify it when reloaded */
-    if (contikiCode == null) {
+    // Remember code, to verify it when reloaded.
+    if (sourceCode == null) {
       final String code = StringUtils.loadFromFile(codeFile);
       if (code != null) {
         String[] lines = code.split("\n");
         if (lineNr-1 < lines.length) {
-          contikiCode = lines[lineNr-1].trim();
+          sourceCode = lines[lineNr-1].trim();
         }
       }
     }
@@ -171,104 +87,5 @@ public class MspBreakpoint implements Watchpoint {
   @Override
   public void removed() {
     mspMote.getCPU().removeWatchPoint((int) address, memoryMonitor);
-  }
-
-  @Override
-  public Collection<Element> getConfigXML() {
-    ArrayList<Element> config = new ArrayList<>();
-    Element element;
-
-    element = new Element("stops");
-    element.setText(String.valueOf(stopsSimulation));
-    config.add(element);
-
-    element = new Element("codefile");
-    File file = mspMote.getSimulation().getCooja().createPortablePath(codeFile);
-    element.setText(file.getPath().replaceAll("\\\\", "/"));
-    config.add(element);
-
-    element = new Element("line");
-    element.setText(String.valueOf(lineNr));
-    config.add(element);
-
-    if (contikiCode != null) {
-      element = new Element("contikicode");
-      element.setText(contikiCode);
-      config.add(element);
-    }
-
-    if (msg != null) {
-      element = new Element("msg");
-      element.setText(msg);
-      config.add(element);
-    }
-
-    if (color != null) {
-      element = new Element("color");
-      element.setText(String.valueOf(color.getRGB()));
-      config.add(element);
-    }
-
-    return config;
-  }
-
-  @Override
-  public boolean setConfigXML(Collection<Element> configXML) {
-    // Already knows mote and breakpoints.
-    for (Element element : configXML) {
-      if (element.getName().equals("codefile")) {
-        var file = mspMote.getSimulation().getCooja().restorePortablePath(new File(element.getText()));
-
-        try {
-          codeFile = file.getCanonicalFile();
-        } catch (IOException e) {
-        }
-
-        if (codeFile == null || !codeFile.exists()) {
-          logger.error("Could not find file: {}", element.getText());
-          return false;
-        }
-      } else if (element.getName().equals("line")) {
-        lineNr = Integer.parseInt(element.getText());
-      } else if (element.getName().equals("contikicode")) {
-        String lastContikiCode = element.getText().trim();
-
-        /* Verify that Contiki code did not change */
-        final String code = StringUtils.loadFromFile(codeFile);
-        if (code != null) {
-          String[] lines = code.split("\n");
-          if (lineNr-1 < lines.length) {
-            contikiCode = lines[lineNr-1].trim();
-          }
-        }
-
-        if (!lastContikiCode.equals(contikiCode)) {
-          logger.warn("Detected modified Contiki code at breakpoint: " + codeFile.getPath() + ":" + lineNr + ".");
-          logger.warn("From: '" + lastContikiCode + "'");
-          logger.warn("  To: '" + contikiCode + "'");
-        }
-      } else if (element.getName().equals("msg")) {
-        msg = element.getText();
-      } else if (element.getName().equals("color")) {
-        color = new Color(Integer.parseInt(element.getText()));
-      } else if (element.getName().equals("stops")) {
-        stopsSimulation = Boolean.parseBoolean(element.getText());
-      }
-    }
-
-    /* Update executable address */
-    address = mspMote.getType().getExecutableAddressOf(codeFile, lineNr);
-    if (address < 0) {
-      logger.fatal("Could not restore breakpoint, did source code change?");
-      return false;
-    }
-    createMonitor();
-
-    return true;
-  }
-
-  @Override
-  public String toString() {
-    return getMote() + ": " + getDescription();
   }
 }
